requirements.txt
requests-2.10.0
django-1.9.6
